* Some clean up * Implement basic support for multiple textareas per stencil in the text tool
svn path=/trunk/koffice/; revision=275738

+++ kivio/kiviopart/kiviosdk/kivio_point.cpp
@@ -69,7 +69,7 @@ KivioPoint::KivioPoint( const KivioPoint &copy )
  *
  * Creates a new point with values.
  */
-KivioPoint::KivioPoint( float newX, float newY, KivioPointType pt )
+KivioPoint::KivioPoint( double newX, double newY, KivioPointType pt )
 {
     m_x = newX;
     m_y = newY;
